Refactor shopping cart popup into file-scoped components
