+++ src/components/shopping-cart/shopping-cart-popover/cart-list.js
@@ -1,0 +1,8 @@
+import { List } from 'antd'
+
+export const CartList = () => {
+    return (
+        <List
+        />
+    )
+}
