+++ src/components/shopping-cart/shopping-cart-popover/index.js
@@ -1,0 +1,1 @@
+export * from './shopping-cart-popover'
