+++ src/components/shopping-cart/shopping-cart-popover/shopping-cart-popover.css
